kubernetes-admin/pyproject.toml
[project]
name = "kubernetes-admin"
version = "0.1.0"
description = ""
authors = [
    {name = "Mathieu GOULIN",email = "mgoulin@mgen.fr"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
